fix(stores): inline pilotStore selector hooks to break circular import
Same selectors-import-store cycle that PR #438 fixed for useGameplayStore
existed in usePilotStore. Inlined `useFilteredPilots` + `usePilotById` at
the bottom of usePilotStore.ts (where they can call the store hook
directly without crossing the module boundary), then deleted the empty
.selectors.ts file. Public API unchanged.

madge: 1 cycle -> 0.
tsc: clean.
jest: 40/40 pilotStore tests pass.
